+++ solvis_graphql_api/composite_solution/composite_rupture_detail.py
@@ -129,15 +129,15 @@ class CompositeRuptureDetail(graphene.ObjectType):
         rupt = rupture_detail(root.model_id, root.fault_system, root.rupture_index)
         # rupt['rate_weighted_mean'] = rupt.map('${:,.2f}'.format)
         # print('RRR', rupt['rate_weighted_mean'].map('{:,.2e}'.format))
-        return round(float(rupt['rate_weighted_mean']), 9)
+        return float(rupt['rate_weighted_mean'])
 
     def resolve_rate_max(root, info, *args, **kwargs):
         rupt = rupture_detail(root.model_id, root.fault_system, root.rupture_index)
-        return round(float(rupt['rate_max']), 9)
+        return float(rupt['rate_max'])
 
     def resolve_rate_min(root, info, *args, **kwargs):
         rupt = rupture_detail(root.model_id, root.fault_system, root.rupture_index)
-        return round(float(rupt['rate_min']), 9)
+        return float(rupt['rate_min'])
 
     def resolve_rate_count(root, info, *args, **kwargs):
         with pd.option_context('display.float_format', '${:,.5f}'.format):
@@ -198,7 +198,7 @@ class FilterRupturesArgs(graphene.InputObjectType):
 
     fault_system = graphene.String(
         required=True,
-        description="One or more fault systems to consider from [`HIK`, `PUY`, `CRU`]",
+        description="The fault systems [`HIK`, `PUY`, `CRU`]",
     )
 
     location_ids = graphene.List(