+++ tests/test_solvis_graphql_api.py
@@ -8,6 +8,7 @@ from graphene.test import Client
 
 from solvis_graphql_api.schema import schema_root
 from solvis_graphql_api.solvis_graphql_api import create_app
+import solvis_graphql_api
 
 
 class TestFlaskApp(unittest.TestCase):
@@ -38,6 +39,18 @@ class TestSchemaAboutResolver(unittest.TestCase):
         executed = self.client.execute(QUERY)
         print(executed)
         self.assertTrue('Hello World' in executed['data']['about'])
+
+    def test_get_about_has_version(self):
+
+        QUERY = """
+        query {
+            about
+        }
+        """
+
+        executed = self.client.execute(QUERY)
+        print(executed)
+        self.assertTrue(solvis_graphql_api.__version__ in executed['data']['about'])
 
 
 class TestSetup(unittest.TestCase):