+++ tests/test_solvis_solution_faults_api.py
@@ -12,6 +12,8 @@ from graphene.test import Client
 import solvis_graphql_api.solution_schema
 from solvis_graphql_api.schema import schema_root  # , matched_rupture_sections_gdf
 
+import pytest
+
 
 def mock_dataframe(*args, **kwargs):
     with open(Path(Path(__file__).parent, 'fixtures', 'geojson.json'), 'r') as geojson:
@@ -47,10 +49,11 @@ QUERY = """
 """
 
 
+@pytest.mark.skip('WIP')
 @mock.patch('solvis_graphql_api.solution_schema.matched_rupture_sections_gdf', side_effect=mock_dataframe)
 class TestSolutionFaultsResolver(unittest.TestCase):
     """
-    A resolver returns info abut faults in a given solution inversion (via SolvisStore.
+    A resolver returns info about faults in a given solution inversion (via SolvisStore.
     """
 
     def setUp(self):
@@ -106,6 +109,7 @@ class TestSolutionFaultsResolver(unittest.TestCase):
         self.assertTrue(gj['features'][0]['properties']['id'] == '5')
 
 
+@pytest.mark.skip('WIP')
 @mock.patch('solvis_graphql_api.solution_schema.matched_rupture_sections_gdf', side_effect=mock_dataframe)
 class TestSolutionLocationsResolver(unittest.TestCase):
     """
@@ -182,6 +186,7 @@ class TestSolutionLocationsResolver(unittest.TestCase):
         )
 
 
+@pytest.mark.skip('WIP')
 class TestSolutionFaultsResolverExceptions(unittest.TestCase):
     """
     A resolver returns info abut faults in a given solution inversion (via SolvisStore.